move SmallJsonTests to src-2 since it seems borked in scala3
