+++ upickle/test/src/upickle/JsonTests.scala
@@ -66,20 +66,6 @@ object JsonTests extends TestSuite {
       """.stripMargin
     val parsed = ujson.read(ugly)
 
-    test("small"){
-      * - {
-        val unparsed = """"\\\uCAFE""""
-        val fromString = ujson.read(unparsed)
-        val fromBytes = ujson.read(unparsed.getBytes)
-        assert(fromString == fromBytes)
-      }
-      * - {
-        val unparsed = """"\/\\\"\uCAFE\uBABE\uAB98\uFCDE\ubcda\uef4A\b\f\n\r\t`1~!@#$%^&*()_+-=[]{}|;:',./<>?""""
-        val fromString = ujson.read(unparsed)
-        val fromBytes = ujson.read(unparsed.getBytes)
-        assert(fromString == fromBytes)
-      }
-    }
     test("correctness"){
       val unparsed = ujson.write(parsed)
       val reparsed = ujson.read(unparsed)
